ServerData in shared
